+++ src/main/java/w/m/kurento/rtsp/demo/RtspDemoApplication.java
@@ -1,13 +1,8 @@
 package w.m.kurento.rtsp.demo;
 
-import org.springframework.beans.factory.config.YamlPropertiesFactoryBean;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.context.support.PropertySourcesPlaceholderConfigurer;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.FileSystemResource;
 
 @SpringBootApplication
 @Configuration
@@ -17,14 +12,5 @@ public class RtspDemoApplication {
 		SpringApplication.run(RtspDemoApplication.class, args);
 	}
 
-	@Bean
-	public static PropertySourcesPlaceholderConfigurer properties() {
-		PropertySourcesPlaceholderConfigurer configurer = new PropertySourcesPlaceholderConfigurer();
-		YamlPropertiesFactoryBean yaml = new YamlPropertiesFactoryBean();
-		//class引入
-		yaml.setResources(new ClassPathResource("videos.yml"));
-		configurer.setProperties(yaml.getObject());
-		return configurer;
-	}
 
 }
